set max width on site intro
to set the max width, it was initially affecting other elements on index. I realised then that the whole body was set to flex display.
I created a new section for the site intro so the width could be styled separately to the case study cards. There is now a max-width on that paragraph of 1050px for readability